Implemented multiple changes - Created routes for login and register - Login and Register are now separate components - The entire logic for handling login is now in AuthContextProvider - Login error is handled through state (not display: none) - Auth code is now in services/ instead of utils/ - Initializing axios inside services/http.js with hard coded base url - Removed proxy from package.json
