+++ package.json
@@ -18,6 +18,7 @@
     "react-dom": "^17.0.2",
     "react-icons": "^4.3.1",
     "react-router": "^6.2.1",
+    "react-router-dom": "^5.3.0",
     "react-scripts": "5.0.0",
     "web-vitals": "^2.1.2"
   },
@@ -57,6 +58,5 @@
     "eslint-plugin-testing-library": "^5.0.1",
     "prettier": "^2.5.1",
     "typescript": "^4.5.4"
-  },
-  "proxy": "http://localhost:1337/"
+  }
 }
