+++ src/services/http.js
@@ -1,0 +1,9 @@
+import axios from 'axios';
+
+const baseURL = 'http://localhost:1337';
+const backendClient = axios.create({
+  baseURL,
+  timeout: 1000,
+});
+
+export default backendClient;
